+++ tests/test_authentication.py
@@ -19,7 +19,7 @@ from tiled.adapters.array import ArrayAdapter
 from tiled.adapters.mapping import MapAdapter
 from tiled.client import Context, from_context
 from tiled.client.auth import CannotRefreshAuthentication
-from tiled.client.context import PasswordRejected
+from tiled.client.context import PasswordRejected, password_grant
 from tiled.server import authentication
 from tiled.server.app import build_app_from_config
 
@@ -780,3 +780,17 @@ def test_admin_can_self_revoke(enter_username_password, config):
             build_app_from_config(config), api_key=api_key["secret"]
         ) as api_context:
             api_context.whoami()
+
+            
+@pytest.mark.timeout(5)
+def test_remaining_authenticated_when_tokens_not_remembered(config):
+    with Context.from_app(build_app_from_config(config)) as context:
+        provider = context.server_info.authentication.providers[0]
+        auth_endpoint = provider.links["auth_endpoint"]
+        tokens = password_grant(
+            context.http_client, auth_endpoint, provider.provider, "alice", "secret1"
+        )
+        context.configure_auth(tokens, remember_me=False)
+        # This will timeout if waiting for a prompt response
+        client = from_context(context, remember_me=False)
+        assert client.context.authenticated